Use the new Eigen typedefs
Commands like the following help in changing all occurrences of a word at once.
  sed -i 's/Matrix4Xd/Matrix4Xf/g' $(rg -l 'Matrix4Xd')

Added two cases for the unit test thresholds, one for double and one for float.

Looks like it is not possible to accurately compute the numerical Jacobians in the tests when using
floats. However since their only purpose is to test that the analytical Jacobians computed in the
library are correct this shouldn't be a problem. The analytical Jacobians match the numerical ones
when using doubles, so we can be certain they are correct. Moreover, the values of the analytical
Jacobians computed with float look reasonable.

+++ include/srl/projection/DistortionBase.hpp
@@ -65,10 +65,10 @@ class DistortionBase
   /// \brief set the generic parameters
   /// @param[in] parameters Parameter vector -- length must correspond numDistortionIntrinsics().
   /// @return    True if the requirements were followed.
-  virtual bool setParameters(const Eigen::VectorXd & parameters) = 0;
+  virtual bool setParameters(const VectorXf & parameters) = 0;
 
   /// \brief Obtain the generic parameters.
-  virtual bool getParameters(Eigen::VectorXd & parameters) const = 0;
+  virtual bool getParameters(VectorXf & parameters) const = 0;
 
   /// \brief The derived class type.
   virtual std::string type() const = 0;
@@ -85,8 +85,8 @@ class DistortionBase
   /// @param[in]  pointUndistorted The undistorted normalised (!) image point.
   /// @param[out] pointDistorted   The distorted normalised (!) image point.
   /// @return     True on success (no singularity)
-  virtual bool distort(const Eigen::Vector2d & pointUndistorted,
-                       Eigen::Vector2d * pointDistorted) const = 0;
+  virtual bool distort(const Vector2f & pointUndistorted,
+                       Vector2f * pointDistorted) const = 0;
 
   /// \brief Distortion and Jacobians.
   /// @param[in]  pointUndistorted  The undistorted normalised (!) image point.
@@ -94,10 +94,10 @@ class DistortionBase
   /// @param[out] pointJacobian     The Jacobian w.r.t. changes on the image point.
   /// @param[out] parameterJacobian The Jacobian w.r.t. changes on the intrinsics vector.
   /// @return     True on success (no singularity)
-  virtual bool distort(const Eigen::Vector2d & pointUndistorted,
-                       Eigen::Vector2d * pointDistorted,
-                       Eigen::Matrix2d * pointJacobian,
-                       Eigen::Matrix2Xd * parameterJacobian = nullptr) const = 0;
+  virtual bool distort(const Vector2f & pointUndistorted,
+                       Vector2f * pointDistorted,
+                       Matrix2f * pointJacobian,
+                       Matrix2Xf * parameterJacobian = nullptr) const = 0;
 
   /// \brief Distortion and Jacobians using external distortion intrinsics parameters.
   /// @param[in]  pointUndistorted  The undistorted normalised (!) image point.
@@ -107,10 +107,10 @@ class DistortionBase
   /// @param[out] parameterJacobian The Jacobian w.r.t. changes on the intrinsics vector.
   /// @return     True on success (no singularity)
   virtual bool distortWithExternalParameters(
-      const Eigen::Vector2d & pointUndistorted,
-      const Eigen::VectorXd & parameters, Eigen::Vector2d * pointDistorted,
-      Eigen::Matrix2d * pointJacobian = nullptr,
-      Eigen::Matrix2Xd * parameterJacobian = nullptr) const = 0;
+      const Vector2f & pointUndistorted,
+      const VectorXf & parameters, Vector2f * pointDistorted,
+      Matrix2f * pointJacobian = nullptr,
+      Matrix2Xf * parameterJacobian = nullptr) const = 0;
   /// @}
 
   //////////////////////////////////////////////////////////////
@@ -121,17 +121,17 @@ class DistortionBase
   /// @param[in]  pointDistorted   The distorted normalised (!) image point.
   /// @param[out] pointUndistorted The undistorted normalised (!) image point.
   /// @return     True on success (no singularity)
-  virtual bool undistort(const Eigen::Vector2d & pointDistorted,
-                         Eigen::Vector2d * pointUndistorted) const = 0;
+  virtual bool undistort(const Vector2f & pointDistorted,
+                         Vector2f * pointUndistorted) const = 0;
 
   /// \brief Undistortion only
   /// @param[in]  pointDistorted   The distorted normalised (!) image point.
   /// @param[out] pointUndistorted The undistorted normalised (!) image point.
   /// @param[out] pointJacobian    The Jacobian w.r.t. changes on the image point.
   /// @return     True on success (no singularity)
-  virtual bool undistort(const Eigen::Vector2d & pointDistorted,
-                         Eigen::Vector2d * pointUndistorted,
-                         Eigen::Matrix2d * pointJacobian) const = 0;
+  virtual bool undistort(const Vector2f & pointDistorted,
+                         Vector2f * pointUndistorted,
+                         Matrix2f * pointJacobian) const = 0;
   /// @}
 
   EIGEN_MAKE_ALIGNED_OPERATOR_NEW
