+++ include/srl/projection/NoDistortion.hpp
@@ -67,14 +67,14 @@ class NoDistortion : public DistortionBase
   /// \brief set the generic parameters
   /// @param[in] parameters Parameter vector -- length must correspond numDistortionIntrinsics().
   /// @return    True if the requirements were followed.
-  bool setParameters(const Eigen::VectorXd & parameters)
+  bool setParameters(const VectorXf & parameters)
   {
     (void)parameters;
     return true;
   }
 
   /// \brief Obtain the generic parameters.
-  bool getParameters(Eigen::VectorXd & parameters) const
+  bool getParameters(VectorXf & parameters) const
   {
     parameters.resize(0);
     return true;
@@ -112,8 +112,8 @@ class NoDistortion : public DistortionBase
   /// @param[in]  pointUndistorted The undistorted normalised (!) image point.
   /// @param[out] pointDistorted   The distorted normalised (!) image point.
   /// @return     True on success (no singularity)
-  bool distort(const Eigen::Vector2d & pointUndistorted,
-                       Eigen::Vector2d * pointDistorted) const
+  bool distort(const Vector2f & pointUndistorted,
+                       Vector2f * pointDistorted) const
   {
     *pointDistorted = pointUndistorted;
     return true;
@@ -125,13 +125,13 @@ class NoDistortion : public DistortionBase
   /// @param[out] pointJacobian     The Jacobian w.r.t. changes on the image point.
   /// @param[out] parameterJacobian The Jacobian w.r.t. changes on the intrinsics vector.
   /// @return     True on success (no singularity)
-  bool distort(const Eigen::Vector2d & pointUndistorted,
-                       Eigen::Vector2d * pointDistorted,
-                       Eigen::Matrix2d * pointJacobian,
-                       Eigen::Matrix2Xd * parameterJacobian = nullptr) const
+  bool distort(const Vector2f & pointUndistorted,
+                       Vector2f * pointDistorted,
+                       Matrix2f * pointJacobian,
+                       Matrix2Xf * parameterJacobian = nullptr) const
   {
     *pointDistorted = pointUndistorted;
-    *pointJacobian = Eigen::Matrix2d::Identity();
+    *pointJacobian = Matrix2f::Identity();
     if (parameterJacobian) {
       parameterJacobian->resize(2, 0);
     }
@@ -146,15 +146,15 @@ class NoDistortion : public DistortionBase
   /// @param[out] parameterJacobian The Jacobian w.r.t. changes on the intrinsics vector.
   /// @return     True on success (no singularity)
   bool distortWithExternalParameters(
-      const Eigen::Vector2d & pointUndistorted,
-      const Eigen::VectorXd & parameters, Eigen::Vector2d * pointDistorted,
-      Eigen::Matrix2d * pointJacobian = nullptr,
-      Eigen::Matrix2Xd * parameterJacobian = nullptr) const
+      const Vector2f & pointUndistorted,
+      const VectorXf & parameters, Vector2f * pointDistorted,
+      Matrix2f * pointJacobian = nullptr,
+      Matrix2Xf * parameterJacobian = nullptr) const
   {
     (void)parameters;
     *pointDistorted = pointUndistorted;
     if (pointJacobian) {
-      *pointJacobian = Eigen::Matrix2d::Identity();
+      *pointJacobian = Matrix2f::Identity();
     }
     if (parameterJacobian) {
       parameterJacobian->resize(2, 0);
@@ -171,8 +171,8 @@ class NoDistortion : public DistortionBase
   /// @param[in]  pointDistorted   The distorted normalised (!) image point.
   /// @param[out] pointUndistorted The undistorted normalised (!) image point.
   /// @return     True on success (no singularity)
-  bool undistort(const Eigen::Vector2d & pointDistorted,
-                 Eigen::Vector2d * pointUndistorted) const
+  bool undistort(const Vector2f & pointDistorted,
+                 Vector2f * pointUndistorted) const
   {
     *pointUndistorted = pointDistorted;
     return true;
@@ -183,12 +183,12 @@ class NoDistortion : public DistortionBase
   /// @param[out] pointUndistorted The undistorted normalised (!) image point.
   /// @param[out] pointJacobian    The Jacobian w.r.t. changes on the image point.
   /// @return     True on success (no singularity)
-  bool undistort(const Eigen::Vector2d & pointDistorted,
-                         Eigen::Vector2d * pointUndistorted,
-                         Eigen::Matrix2d * pointJacobian) const
+  bool undistort(const Vector2f & pointDistorted,
+                         Vector2f * pointUndistorted,
+                         Matrix2f * pointJacobian) const
   {
     *pointUndistorted = pointDistorted;
-    *pointJacobian = Eigen::Matrix2d::Identity();
+    *pointJacobian = Matrix2f::Identity();
     return true;
   }
   /// @}
